requirements.txt
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
scikit-surprise==1.1.1